api/requirements.txt
fastapi
google-generativeai
supabase
uvicorn
pydantic
python-dotenv